Commit Window: refactor
We set the disclosure button's action to `toggleTableView:` to show/hide the tableView. It will be easier to do animation in this method instead of the setter for the `showsTableView` property (via bindings), since we may not want to always animate the table when we set the `showsTableView` property (e.g., when first showing the commit window).

Also simplified the auto layout constraints.