Show both display name and path in file/favorite chooser
Ranking has also been simplified in the file chooser: by default we match against the full path but prefer matches that only touch the base name.

The current document is placed last when there is no filter string (where we LRU sort) but once there is a filter string, we do not give it special treatment.