Ensure valid match/excerpt ranges for CRLF delimited files
When searching for either CR or LF in a CRLF delimited file, we effectively match a partial newline and previously the excerpt’s BOL/EOL positions would not take that into account, so the match wasn’t guaranteed to be a subset of the excerpt.

Ideally the document reader should harmonize newlines so that the searcher only sees LF, though currently that is not the case.