Improve bundle item chooser
We now show two lines per item: the name of the item and the location on second line.

For settings we show the (literal) value of the item next to its name (ideally user could click the item to toggle between literal and effective value).

Selecting settings items (from properties files) will go to the proper line in the properties file.