Fix potential crash in file chooser
This was caused by performing regexp matching on a temporary string object and later accessing captures from the match (when the string object was out of scope).

Rather than extend the scope of the string object I rewrote the code to use Cocoa’s NSRegularExpression class so we avoid using C++ (and thus temporary objects).