Reorder accessibility order of chooser cell view items
When VoiceOver users want to efficiently find an item in a list, they often go
through the items quickly, hearing just first few characters to decide whether
it is the item they want or whether it is positively not the one (in which case
they move to the next item). Therefore it is necessary for each row (item) of a
table to be worded in such a way for VoiceOver so that the most important piece
of information is read first, then followed by the less important piece of
information, etc.

E.g. for the file chooser, the most important piece of information about a file
is its file name. So we put that first for the VoiceOver user. If they are
unsure whether it is the right one (perhaps because there is more than one file
with the same name in different directories), then path to the file is read
after that.

Fixes textmate/bugs#10