Macro recording would not replay inserted text
This broke when we changed from insertText: → insertText:replacementRange: as the latter selector is not replayable.

We now record this action simply as insertText: and ignore the “replacementRange” argument. This argument can be used by advanced input managers, so there might be an issue with text recorded using such manager, however, to properly support this we need to translate the “replacementRange” into something relative, which is left as an exercise for a rainy day.