Do not invoke “did save” callback for untitled documents
The problem is that when something is saved, the text view is responsible for invoking the callbacks, but it might not show the document that was saved (e.g. in the case of Save All where background tabs are saved).

Long-term this should be improved so that OakTextView is not responsible for executing these commands; the challenge is that OakTextView supply a lot of the environment (like some of the TM variables), potentially the input, and is also the potentially receiver of the output.