Convert filter string to decomposed unicode
Since we are doing a literal (byte-by-byte) subset match we need to have the filter string use the same unicode normalization form as the items it is matched against.

For favorites and the file chooser, the items are file system names, which default to the decomposed form.

Ideally the filtering algorithm would know about diacritics, but in practice the current solution should work for most actual scenarios, and is much simpler (thus faster).