Support reloading tab bar view while a tab is being dragged
This would happen if dragging a tab into an unmodified document (in the same window) while holding down control (to insert the path of the dragged tab).

Here the document would change state to “modified” which would trigger a tab bar view reload, but since the drag session isn’t completed, we reload the tabs while one of them is “outside” the bar, which causes it to be added, and once the drag session completes, the original tab outside the bar, will be re-added, leading to two copies of the same tab.