Refactor the file browser
We now subclass NSOutlineView to detect recursive expansion/collapsing rather than look at the current event (and testing for option). This should be more reliable, for example previously using drag’n’drop with option could falsely trigger recursive expansion.

The way background folder content is loaded is also made simpler with a fairly straightforward code flow (thanks in part to the use of blocks) without the need for “pendingXYZ” variables that got checked at somewhat arbitrary times (to see if we needed to expand/select items that got loaded).

This also better tracks expanded items, as we now remove items that are collapsed from the session info the minute they are collapsed, rather than take a snapshot of the outline view during application termination. This should close #1111.

Another improvement is that we wrap item updating in begin/end update calls to get animated expansion of folders.

There are still a few issues, for example switching to using the view-based API for the outline view and handling reloading items while the field editor is up (which I think require the former).

Also, this commit does not change the data source API so there is some redundant work being done, as we request a reload after getting a “did reload” notification.