Remove file browser’s code for return/backtab in field editor
Handling return was to prevent that key to move to editing the next row, though it does not seem to have this behavior anymore.

Backtab was to support editing the row above, though Apple has a defined behavior for how tab/backtab should work when editing table view items, so we shouldn’t change that.