Delete code that tests if field editor was dismissed
It seems that after the field editor has been dismissed, the first responder reverts to the outline view (file browser), even though we are asked if we accept first responder.

So now we simply check if we are already first responder, and that seems to work just as well (for clicking on an outline view row while the field editor is showing).