Don’t pass NSTableCellView to OakFileManager
The view passed to the file manager is used to request the undo manager. The problem with using an NSTableCellView instance is that this view may be taken out of the view hierarchy (and thus responder chain) which makes it lose its undo manager.

This isn’t an immediate problem but if we later use undo/redo, it might be.