Add support for highlighting lines with marks
To set the color of the highlight:

1. Add a new setting to the theme with the scope
"mark.inline.<mark_type>", where <mark_type> is the mark type to be
highlighted.

2. Add the theme component, "highlight", to set the highlight color.

The following setting will highlight marks with the mark type "error"
with a light red color. The mark type is used when setting a mark
with the "mate" command: "mate -s error:foo".

{ name = 'Mark inline: error';
  scope = 'mark.inline.error';
  settings = {
    highlight = '#FFD3c5';
  };
}