Prevent opening two tabs for the same file
This could previously be done by creating a symbolic link and then selecting both the symbolic link and the original file, and selecting to open both at once.

Since the two tabs would have same identifier, the NSView for one of the two tabs could be left behind in the tab bar when manipulating the tabs.