Change Find/HTML window collection behavior to work with 10.11
Previously we would set the window to “join all spaces” when TextMate was active, but on 10.11 it seems that the system will act like this is set even when TextMate “becomes activate”, so moving focus to TextMate, with an open Find/HTML window, will move the window to the active space, rather than move to the space containing TextMate.

Now we only set “join all spaces” when the user requests one of the windows, and the window is currently not already on the active space, and we unset this mode after bringing the window to front.

This commit effectively reverts a52fc3493bc8b27556828fc30c5e385702a79ed8