Prefix some of the menu items in Navigate with "Jump to"
The Navigate menu now contains a mixture of items with prefixes "Jump to", "Go to" or none at all. The "Go to" prefix is especially inconsistent since we have already have a dedicate "Go" item in the Main menu bar. For consistency, change them to use the "Jump to" prefix.

The commit also renames the appropriate dialog boxes and updates some documentation to refer to the new names.