Deleting last item in pasteboard history browser is now a no-op
Unfortunately disabling the “Delete” button is cumbersome because our history is stored with CoreData so knowing the number of items in the history requires a fetch request, which we would not know when to re-run (to update the disabled state when the history is changed).

When the filter string is empty we can use the array controller’s results, but I’m leaving that for the future.