Change FFResultNode’s ‘ignore’ property to ‘readOnly’
At the same time the semantics have been simplified so that the ‘readOnly’ property (formerly ‘ignored’) does not have any side effects with respect to whether or not the item is excluded.

The only thing special is that when a node is set to be read-only, a change of its ‘exclude’ property will not be forwarded to its children.