Limit the steps taken by OakDocument to find a file type
If there is no custom binding and no installed grammar matches the document (based on content or path extension) then we leave the document without a file type and expect the caller (of loadModalForWindow:completionHandler:) to set the file type (in the completion handler).

The caller can consult online bundles, present UI to the user, and also know about “project directory”, so they can make a better choice than the OakDocument.