Use default monospace font when failing to get requested font
If we ask for a non-existing font then we may get back the system default font and as most users expect a monospaced font, we take extra steps to catch this situation and switch to the default monospaced font.

For example users that are currently using Menlo may get into this situation on macOS 10.12 where Menlo has been removed.

Furthermore, when no font name is configured we also use the system’s default fixed width font.