Move clipboard wrapper to (new) OakDocument framework
Ideally we would stay in the document framework, but the editor framework currently depends on our document framework and we want to introduce a OakDocumentEditor which encapsulates an editor and layout instance, so to avoid cyclic dependencies we are doing this as a new framework.

Once the last document dependencies have been removed from the editor framework, we can merge OakDocument and OakDocumentEditor.