Redirect protocol-relative URLs in HTML output to https
The problem is that HTML output is loaded from the x-txmt-command protocol, possibly with the “base URL” set to file://$TM_FILEPATH, so protocol-relative URLs like //example.org/main.css would be resolved as file://example.org/main.css.

We now try to detect these URLs (heuristically, since we only see the resolved URL) and (by default) redirect to https.

This commit checks for the presence of NSURLComponents which was introduced in Mac OS X 10.9. On 10.8 this commit has no effect.