Introduce exit_insert_snippet_no_indent
By default a command with snippet output will have the snippet indented to the current indent level.

This can be disabled in the bundle item’s property list but there might be situations where we want to disable it dynamically.

For an example see http://lists.macromates.com/textmate/2016-August/039845.html