Handle begin/end undo grouping for document w/o document editor
Previously this would be a no-op basically bringing the undo stack out of sync and making TextMate prone to crashing.

We create an undo group when reloading document changes and here we do not create a document editor, hence when this happened for unselected tabs (which currently have no document editor), the undo stack was brought out of sync.