Let file name formatter return error string for unexpected value type
One spin report shows that an attributed string is being used with the text field. Calling ‘description’ on an attributed string can produce a lot of output (dumping all the attributes) and the spin report in question shows a lot of symbols related to auto-layout and text sizing, so perhaps this is related.

http://lists.macromates.com/textmate/2016-August/039891.html