Re-introduce the use of bindings for file browser’s text field
The display name of an item may change, e.g. when user toggles “hidden extension” in Finder. The data source for the FSItem will catch this and update the displayName property, but since we did not observe it, the file browser would not update with the new name.

The reason we disabled use of bindings was because the binding is bidirectional and we had no setter for the new object value.

We solve this by adding a new property to FSItem that is for use with bindings and data formatters.