Remove code that was never executed
Since we were testing for a class rather than instance method, this check would never be true.

Since I am not aware of any problems with the current code where we do not call super, I decided to remove rather than change the code (to call super when it responds to the selector).