Use the table view as Open Recent Project’s initial first responder
If the user starts typing the focus is moved to the search field, so the only potential disadvantage is the discoverability.

The advantage is that with the table view as initial first responder we can support forward and backward delete to remove items, where previously we required ⌘⌫, which would work when editing text, and thus could cause data loss for a user expecting this key to instead delete to the beginning of the line (text field).