Use OakDocumentController to track OakDocument instances
Currently all OakDocument factory methods goes via the document controller but it would probably be better to move these methods to the document controller so that all document creation happens via the controller (unless one wants to opt out of the services provided, e.g. documents used with the bundle editor and commit window do not need to be “tracked”).