Let the file chooser work with OakDocument instead of document_t
There is a noticeable performance degradation but this is only because when we obtain an OakDocument from a document_t object then an observer is added to OakDocument to keep the document_t object in sync.

Once we have directory enumeration code that return OakDocument instances, this problem will disappear.