Add mutex to make code (sort of) safe to use from multiple threads
The add/remove functions are called when opening/closing documents, in TextMate this happens in the main thread, but not when executing the tests. With the mutex, it should now be safe to call the functions from multiple threads at the same time although if file system changes are actually observed, the callback will be executed on the main thread.